add simple extension, update tests, move previous simple into webpacked extension
