+++ extensions/README.md
@@ -1,3 +1,3 @@
-This directory is for local extensions only.
+Extensions are client-side additions for Genetic Constructor, which allow for interaction with user's data from the client, and widgets inserted at specified regions within the app.
 
-Extensions that are accessible to the server should be listed in `/server/extensions/package.json` and will be installed using npm. Extensions in this folder will only be accessible if listed in that manifest.
+This directory is for local extensions only. They will not automatically be installed. Extensions that are accessible to the server should be listed in `/server/extensions/package.json` and will be installed using npm. Extensions in this folder will only be accessible if listed in that manifest.
