+++ extensions/simple/README.md
@@ -1,9 +1,9 @@
-Very simple example plugin
+A Simple Genetic Construct Extension
 
-## Installation
+This extension is only a single file, which will log events as they happen in the app.
 
-You will need to install `webpack`
+It demonstrates what is minimally needed to write an extension, and demonstrates how to register it with the App.
 
-`npm install`
+This extension does not have any visual element - it simply exists and interacts with the app non-visually.
 
-`webpack main.js index.js`
+No building or installation is necessary.
