+++ extensions/webpacked/README.md
@@ -1,0 +1,11 @@
+Example plugin which uses webpack and displays in the sequenceDetail region.
+
+Webpack is necessary for bundling larger projects, as only one file (`index.js`) is allowed to be served per extension.
+
+## Installation
+
+You will need to install `webpack` (https://webpack.github.io/) so you can run it from the command line.
+
+`npm install`
+
+`webpack main.js index.js`
